Move safe_name into lib
