+++ tools/generator_common.py
@@ -1,6 +1,8 @@
 from collections import namedtuple
 from pathlib import Path
 
+from html_compose.util_funcs import safe_name
+
 AttrDefinition = namedtuple(
     "AttrDefinition", ["name", "safe_name", "value_desc", "description"]
 )
@@ -11,22 +13,6 @@ def get_path(fn):
         return Path("tools") / fn
     else:
         return Path(fn)
-
-
-def safe_name(name):
-    """
-    Some names are reserved in Python, so we need to add an underscore
-    An underscore after was chosen so type hints match what user is looking for
-    """
-    # Keywords
-    if name in ("class", "is", "for", "as", "async", "del"):
-        name = name + "_"
-
-    if "-" in name:
-        # Fixes for 'accept-charset' etc.
-        name = name.replace("-", "_")
-
-    return name
 
 
 def ReadAttr(attr_spec) -> AttrDefinition:
